删除模块和功能

+++ src/main/java/com/thf/service/Impl/FunctionServiceImpl.java
@@ -24,55 +24,73 @@ public class FunctionServiceImpl implements FunctionService {
     ProjectDAO projectDAO;
     @Resource
     ModuleDAO moduleDAO;
+
     @Override
-    public ResultVO createFunc(String token,int mid, Function function) {
-        int uid= (int) JwtUtil.parseToken(token).get("id");
-        Project project=new Project();
+    public ResultVO createFunc(String token, int mid, Function function) {
+        int uid = (int) JwtUtil.parseToken(token).get("id");
+        Project project = new Project();
         project.setCreateUser(uid);
-        List<Project> lp=null;
-        List<Module> lm=null;
-        if((lp=projectDAO.getAllProject(project))==null){
-            Res.res(2000,"请先创建项目");
-        }else {
-            for (Project po:lp) {
-                lm=moduleDAO.searchAllModule(po.getProjectId());
-                for (Module mo:lm) {
-                    if(mo.getMId()==mid){
+        List<Project> lp = null;
+        List<Module> lm = null;
+        if ((lp = projectDAO.getAllProject(project)) == null) {
+            Res.res(2000, "请先创建项目");
+        } else {
+            for (Project po : lp) {
+                lm = moduleDAO.searchAllModule(po.getProjectId());
+                for (Module mo : lm) {
+                    if (mo.getMId() == mid) {
                         function.setMId(mid);
                         function.setFCreateTime(System.currentTimeMillis());
-                        if(functionDAO.insert(function)>0){
-                            return Res.res(2000,"新增功能成功",functionDAO.searchById(function.getFId()));
-                        }else {return Res.res(5000,"新增失败");}
+                        if (functionDAO.insert(function) > 0) {
+                            return Res.res(2000, "新增功能成功", functionDAO.searchById(function.getFId()));
+                        } else {
+                            return Res.res(5000, "新增失败");
+                        }
                     }
                 }
-               return Res.res(2000,"请先创建模块");
+                return Res.res(2000, "请先创建模块");
             }
         }
-        return Res.res(5000,"新增失败");
+        return Res.res(5000, "新增失败");
     }
 
     @Override
-    public ResultVO updateFunc(Function function) {
-        if(functionDAO.update(function)>0){
-            return new ResultVO(2000,"修改成功",functionDAO.searchById(function.getFId()));
+    public ResultVO updateFunc(String token, Function function) {
+        int uid = (int) JwtUtil.parseToken(token).get("id");
+        int mpid = moduleDAO.searchBymId(function.getMId()).getPId();
+        if (projectDAO.searchById(mpid).getCreateUser() == uid) {
+            if (functionDAO.update(function) > 0) {
+                return Res.res(2000, "修改成功", functionDAO.searchById(function.getFId()));
+            }
+            return Res.res(5000, "修改失败");
         }
-        return new ResultVO(2000,"修改失败");
+        return Res.res(4000, "仅可更改自己账户下的功能");
     }
 
     @Override
-    public ResultVO deleteFunc(int fid) {
-        if(functionDAO.delete(fid)>0){
-           return Res.res(2000,"删除成功",null);
+    public ResultVO deleteFunc(String token, int[] fidlist) {
+        int uid = (int) JwtUtil.parseToken(token).get("id");
+        for (int fid:fidlist) {
+            int mpid = moduleDAO.searchBymId(fid).getPId();
+            if (projectDAO.searchById(mpid).getCreateUser() == uid) {
+                if (functionDAO.delete(fid) > 0) {
+                    if(fidlist[fidlist.length-1]==fid) {
+                        return Res.res(2000, "删除成功");
+                    }
+                }
+                return Res.res(5000, "删除失败");
+            }
+            return Res.res(4000, "仅可删除自己账户下的功能");
         }
-        return Res.res(2000,"删除失败",null);
+        return Res.res(5000, "服务器内部错误");
     }
 
     @Override
     public ResultVO searchFunc(int fid) {
-        Function f=functionDAO.searchById(fid);
-        if(f!=null){
-            return new ResultVO(2000,"搜索成功",f);
+        Function f = functionDAO.searchById(fid);
+        if (f != null) {
+            return Res.res(2000, "搜索成功", f);
         }
-        return new ResultVO(2000,"搜索失败");
+        return Res.res(2000, "搜索失败");
     }
 }
